Avoid the need to interactivly stop the initial publisher
The publisher was started twice by the sample script:
The first time is was started with the $_HUB_CS as argument, to
register itself at the IoT-Hub.
Then it should be stopped interactivly (with ctrl-c), so that the 2nd
instance is started.
The first instance is not needed, if the 2nd gets also the $_HUB_CS as start argument.

Signed-off-by: Juergen Kosel <juergen.kosel@softing.com>